Split Gauntlet primitive tests into new file
Moves Gauntlet-specific physics primitive tests out of `PhysicsCharts.test.tsx` into a dedicated `GauntletChart.test.tsx`, including spawn placement, negative mark rendering, pop-by-identity behavior, and net core-force direction checks. This keeps the shared physics chart test file focused on chart-level behavior and reduces unnecessary imports there.

Also includes minor formatting cleanup in `GauntletChart.tsx` and `PhysicsKernel.ts` (comment tightening and compact type/interface formatting) with no functional runtime changes.
